refactor code

+++ internal/models/policy.go
@@ -5,7 +5,7 @@ import "time"
 type Policy struct {
 	ID             [16]byte  `json:"id"`
 	ContactPointID [16]byte  `json:"contact_point_id"`
-	Severity       int16     `json:"severity"`
+	Severity       int       `json:"severity"`
 	Status         string    `json:"status"`
 	Action         string    `json:"action"`
 	CreatedAt      time.Time `json:"created_at"`
